requirements.txt
selenium>=4.10.0
webdriver-manager>=3.8.6
playwright
playwright-stealth
Pillow
